+++ MapImporter/ObjectGroup.cs
@@ -58,7 +58,7 @@ namespace MapImporter
         /// The opacity of the layer as a value from 0 to 1.
         /// Defaults to 1.
         /// </summary>
-        public int Opacity { set; get; }
+        public float Opacity { set; get; }
         /// <summary>
         /// Whether the layer is shown (1) or hidden (0). Defaults to 1.
         /// </summary>
@@ -84,7 +84,7 @@ namespace MapImporter
         /// <summary>
         /// Constructor for the ObjectGroup class.
         /// </summary>
-        public ObjectGroup(string name, int x, int y, int width, int height, int opacity, Color color)
+        public ObjectGroup(string name, int x, int y, int width, int height, float opacity, Color color)
         {
             Name = name;
             X = x;
