+++ MapImporter/TileLayer.cs
@@ -35,7 +35,7 @@ namespace MapImporter
         /// <summary>
         /// The opacity of the layer as a value from 0 to 1. Defaults to 1.
         /// </summary>
-        public int Opacity { set; get; }
+        public float Opacity { set; get; }
         /// <summary>
         /// Whether the layer is shown (1) or hidden (0). Defaults to 1.
         /// </summary>
@@ -65,7 +65,7 @@ namespace MapImporter
         /// <param name="width"></param>
         /// <param name="height"></param>
         /// <param name="opacity"></param>
-        public TileLayer(string name, int x, int y, int width, int height, int opacity)
+        public TileLayer(string name, int x, int y, int width, int height, float opacity)
         {
             Name = name;
             X = x;
